+++ src/test/java/net/fabricmc/mappingio/VisitOrderVerifyingVisitor.java
@@ -111,7 +111,7 @@ public class VisitOrderVerifyingVisitor extends ForwardingMappingVisitor {
 
 	@Override
 	public boolean visitContent() throws IOException {
-		assertNamespacesVisited();
+		assertHeaderVisited();
 		assertContentNotVisited();
 
 		visitedContent = true;